+++ packages/common/utils/src/public_api.ts
@@ -1,3 +1,5 @@
+export { serializeXmlToString } from './serialize-xml-to-string';
+export { parseXmlFromString } from './parse-xml-from-string';
 export { checkValueIsEmpty } from './check-value-is-empty';
 export { detectChanges } from './detect-changes';
 export { getBodyRect } from './get-body-rect';