+++ packages/stream/jwt/src/to-base64.ts
@@ -1,0 +1,3 @@
+export function toBase64<T>(pojo: Record<string, unknown> | T): string {
+    return window.btoa(unescape(encodeURIComponent(JSON.stringify(pojo))));
+}