+++ CHANGELOG.md
@@ -1,3 +1,54 @@
+#### 14.32.2 (2020-09-04)
+
+##### Continuous Integration
+
+-   update devDependencies ([#115](https://github.com/Angular-RU/angular-ru-sdk/pull/115))
+    ([c6d70075](https://github.com/Angular-RU/angular-ru-sdk/commit/c6d70075f40f2b702ecfc2782eeea3475691933c))
+
+#### 14.32.1 (2020-09-03)
+
+##### Bug Fixes
+
+-   **@angular-ru/common:** support TypeScript 4
+    ([238a9017](https://github.com/Angular-RU/angular-ru-sdk/commit/238a9017e8579a19bbdfe8a30cbd73edb4b862ee))
+
+##### Refactors
+
+-   **@angular-ru/stream:** remove tsconfig.base.json
+    ([1c4cee80](https://github.com/Angular-RU/angular-ru-sdk/commit/1c4cee8092e0196c67f36c4f3a94e02fe8870a62))
+
+#### 14.32.0 (2020-09-03)
+
+##### Chores
+
+-   **release:** updated release notes and package versions [ci skip]
+    ([5edd7e3c](https://github.com/Angular-RU/angular-ru-sdk/commit/5edd7e3c2479ad28c6f5c17407d0ccc968b0ce71))
+
+##### Continuous Integration
+
+-   update dependency @types/node to v14.6.3
+    ([9e9b2252](https://github.com/Angular-RU/angular-ru-sdk/commit/9e9b2252808998fa7ce893f3e7bc176ee55bf228))
+
+##### New Features
+
+-   **@angular-ru/common:**
+    -   add serializeXmlToString
+        ([ebd60f46](https://github.com/Angular-RU/angular-ru-sdk/commit/ebd60f463ab76d38460244dba40a1e7b8da4cd60))
+    -   add parseXmlFromString
+        ([69a49f86](https://github.com/Angular-RU/angular-ru-sdk/commit/69a49f8683ed493470501fcc2d677ff69858fecb))
+
+#### 14.31.0 (2020-09-03)
+
+##### Continuous Integration
+
+-   update dependency @types/node to v14.6.3
+    ([9e9b2252](https://github.com/Angular-RU/angular-ru-sdk/commit/9e9b2252808998fa7ce893f3e7bc176ee55bf228))
+
+##### New Features
+
+-   **@angular-ru/common:** add parseXmlFromString
+    ([69a49f86](https://github.com/Angular-RU/angular-ru-sdk/commit/69a49f8683ed493470501fcc2d677ff69858fecb))
+
 #### 14.30.1 (2020-09-02)
 
 ##### Continuous Integration